analizador_emociones/requirements.txt
opencv-python
face_recognition
deepface
tf-keras
pandas
numpy
pillow